added cli support for data selection
To make things a bit easier to use via the
CLI I added a very crude "parser" to parse
RangePoints from strings.

Any now for a heads-up...
The way we implemented Poisson Regression
results in multiple (maybe different) edge
cases, I've described them in the docstring
of the predictor class.