fixed funny bug
The funny bug was dew to the end-year
being None which does indeed not make
much sense to compare to a number via
<=...
Additionally I went ahead and change
the way runtime data is stored. It is
now a "external start object", aka.
AppData. Which is a global Instance
of the class _AppData similar to how
DB is handled.

Relates to #6